Create getBg3PakMeta.pl
Script to extract the metadata from BG3 mod .pak files that are needed for manual addition to 'modsettings.lsx'.  The only argument is the path to the module .pak file.  Intended for us poor Mac users, as it should run on a Mac without any other dependencies.  It also doesn't use any externally licensed code, so have at it...